# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DemoAuthInputRole = typeof DemoAuthInputRole[keyof typeof DemoAuthInputRole];


export const DemoAuthInputRole = {
  Student: 'Student',
  Faculty: 'Faculty',
  Club_President: 'Club President',
  Club_Vice_President: 'Club Vice President',
  Administrator: 'Administrator',
} as const;

export interface DemoAuthInput {
  role: DemoAuthInputRole;
}

export interface DemoUser {
  id: number;
  name: string;
  email: string;
  role: string;
  initials: string;
}

export interface DashboardStat {
  label: string;
  value: string;
  detail: string;
  tone: string;
}

export interface Announcement {
  id: number;
  title: string;
  summary: string;
  category: string;
  publishedAt: string;
  source: string;
  isPinned: boolean;
}

export interface CampusEvent {
  id: number;
  title: string;
  date: string;
  time: string;
  location: string;
  category: string;
  color: string;
}

export interface Deadline {
  id: number;
  title: string;
  course: string;
  dueDate: string;
  dueLabel: string;
  status: string;
  priority: string;
}

export interface Dashboard {
  user: DemoUser;
  stats: DashboardStat[];
  announcements: Announcement[];
  events: CampusEvent[];
  deadlines: Deadline[];
}

export interface Club {
  id: number;
  name: string;
  category: string;
  description: string;
  members: number;
  nextMeeting: string;
  accent: string;
}

export interface Opportunity {
  id: number;
  title: string;
  organization: string;
  type: string;
  deadline: string;
  description: string;
  tag: string;
}

export interface Assignment {
  id: number;
  title: string;
  course: string;
  dueDate: string;
  progress: number;
  status: string;
}

export interface Issue {
  id: number;
  title: string;
  category: string;
  location: string;
  description: string;
  status: string;
  createdAt: string;
  reportedBy: string;
}

export interface IssueInput {
  /** @minLength 2 */
  title: string;
  category: string;
  location: string;
  /** @minLength 5 */
  description: string;
}

export type GetAnnouncementsParams = {
q?: string;
};