# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /demo-auth:
    post:
      operationId: demoAuth
      tags: [auth]
      summary: Start a demo session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoAuthInput"
      responses:
        "200":
          description: Demo session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoUser"
...
  /announcements:
    get:
      operationId: getAnnouncements
      tags: [announcements]
      summary: List verified announcements
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
...
  /issues:
    get:
      operationId: getIssues
      tags: [issues]
      summary: List reported campus issues
      responses:
        "200":
          description: Issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
    post:
      operationId: createIssue
      tags: [issues]
      summary: Report a campus issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueInput"
      responses:
        "201":
          description: Created issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
components:
  schemas:
...
    DemoAuthInput:
      type: object
      properties:
        role:
          type: string
          enum: [Student, Faculty, Club President, Club Vice President, Administrator]
      required: [role]
    DemoUser:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        initials: { type: string }
      required: [id, name, email, role, initials]
...
    Announcement:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        summary: { type: string }
        category: { type: string }
        publishedAt: { type: string }
        source: { type: string }
        isPinned: { type: boolean }
      required: [id, title, summary, category, publishedAt, source, isPinned]
...
    Issue:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        category: { type: string }
        location: { type: string }
        description: { type: string }
        status: { type: string }
        createdAt: { type: string }
        reportedBy: { type: string }
      required: [id, title, category, location, description, status, createdAt, reportedBy]
    IssueInput:
      type: object
      properties:
        title: { type: string, minLength: 2 }
        category: { type: string }
        location: { type: string }
        description: { type: string, minLength: 5 }
      required: [title, category, location, description]