# lib/api-zod/src/generated/types/announcement.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Announcement {
  id: number;
  title: string;
  summary: string;
  category: string;
  publishedAt: string;
  source: string;
  isPinned: boolean;
}