# lib/api-zod/src/generated/types/demoAuthInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DemoAuthInputRole } from './demoAuthInputRole';

export interface DemoAuthInput {
  role: DemoAuthInputRole;
}