# lib/api-zod/src/generated/types/demoUser.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DemoUser {
  id: number;
  name: string;
  email: string;
  role: string;
  initials: string;
}