# lib/api-zod/src/generated/types/issue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Issue {
  id: number;
  title: string;
  category: string;
  location: string;
  description: string;
  status: string;
  createdAt: string;
  reportedBy: string;
}