# lib/api-zod/src/generated/types/issueInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface IssueInput {
  /** @minLength 2 */
  title: string;
  category: string;
  location: string;
  /** @minLength 5 */
  description: string;
}